+++ Tests/src/PublicAPITests/AnchorSystemTests.cpp
@@ -103,7 +103,7 @@ void DeleteAnchors(csp::systems::AnchorSystem* AnchorSystem, const csp::common::
 
     if (Result.GetResultCode() == csp::systems::EResultCode::Success)
     {
-        for (int idx = 0; idx < AnchorIDs.Size(); idx++)
+        for (size_t idx = 0; idx < AnchorIDs.Size(); idx++)
         {
             std::cerr << "Anchor Deleted: "
                       << "Id=" << AnchorIDs[idx] << std::endl;
@@ -385,7 +385,7 @@ CSP_PUBLIC_TEST(CSPEngine, AnchorSystemTests, GetAnchorsInsideCircularAreaTest)
         const auto& ResultAnchors = Result.GetAnchors();
         AnchorCollection = csp::common::Array<csp::systems::Anchor>(ResultAnchors.Size());
 
-        for (int idx = 0; idx < ResultAnchors.Size(); ++idx)
+        for (size_t idx = 0; idx < ResultAnchors.Size(); ++idx)
         {
             AnchorCollection[idx] = ResultAnchors[idx];
         }
